+++ NEWS.md
@@ -1,3 +1,14 @@
+# dynamite 1.4.0
+
+  * `dynamite()` now supports parallel computation via the reduce-sum functionality of Stan.
+  * Fixed an issue in `predict()` that resulted in redundant `NAs produced` warnings.
+  * Estimation of dynamic multivariate panel models with multiple imputation is now available via the function `dynamice()` which uses the `mice` package.
+  * Fixed an issue with `formula.dynamitefit()` with models that had multivariate channels.
+
+# dynamite 1.3.3
+
+  * Fixed a partial argument name issue in the internal `update()` method used by `lfo()`.
+
 # dynamite 1.3.2
 
   * Fixed the regularization of the default priors so that they match with the priors vignette.